feat(v2): sign-in + project list UI over typed IPC; token-never-in-bindings guard
AdoError renders as friendly text per variant; RateLimited uses u32 secs
(specta forbids u64 across IPC). Live AAD sign-in check pending user.

+++ v2/src-tauri/src/ado.rs
@@ -16,8 +16,9 @@ pub struct Project {
 pub enum AdoError {
     #[error("unauthorized")]
     Unauthorized,
+    // u32 (not u64): specta forbids BigInt-style types crossing IPC.
     #[error("rate limited, retry after {retry_after_secs}s")]
-    RateLimited { retry_after_secs: u64 },
+    RateLimited { retry_after_secs: u32 },
     #[error("forbidden")]
     Forbidden,
     #[error("not found")]
